+++ examples/with-redux-thunk/components/clock.js
