+++ examples/with-redux-thunk/components/counter.js
@@ -1,22 +1,21 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import { bindActionCreators } from 'redux'
 import { incrementCount, decrementCount, resetCount } from '../store'
 
 class Counter extends Component {
   increment = () => {
-    const { incrementCount } = this.props
-    incrementCount()
+    const { dispatch } = this.props
+    dispatch(incrementCount())
   }
 
   decrement = () => {
-    const { decrementCount } = this.props
-    decrementCount()
+    const { dispatch } = this.props
+    dispatch(decrementCount())
   }
 
   reset = () => {
-    const { resetCount } = this.props
-    resetCount()
+    const { dispatch } = this.props
+    dispatch(resetCount())
   }
 
   render () {
@@ -38,6 +37,5 @@ function mapStateToProps (state) {
   const { count } = state
   return { count }
 }
-const mapDispatchToProps = dispatch => bindActionCreators({ incrementCount, decrementCount, resetCount }, dispatch)
 
-export default connect(mapStateToProps, mapDispatchToProps)(Counter)
+export default connect(mapStateToProps)(Counter)
