+++ examples/with-redux-thunk/components/examples.js
