+++ examples/with-redux-thunk/lib/with-redux-store.js
