+++ examples/with-redux-thunk/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "with-redux",
+  "version": "1.0.0",
+  "scripts": {
+    "dev": "next",
+    "build": "next build",
+    "start": "next start"
+  },
+  "dependencies": {
+    "next": "latest",
+    "react": "^16.7.0",
+    "redux-devtools-extension": "^2.13.2",
+    "react-dom": "^16.7.0",
+    "react-redux": "^5.0.1",
+    "redux": "^3.6.0",
+    "redux-thunk": "^2.1.0"
+  },
+  "license": "ISC"
+}
