+++ examples/with-redux-thunk/pages/_app.js
