+++ examples/with-redux/components/counter.js
@@ -1,21 +1,22 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
 import { incrementCount, decrementCount, resetCount } from '../store'
 
 class Counter extends Component {
   increment = () => {
-    const { dispatch } = this.props
-    dispatch(incrementCount())
+    const { incrementCount } = this.props
+    incrementCount()
   }
 
   decrement = () => {
-    const { dispatch } = this.props
-    dispatch(decrementCount())
+    const { decrementCount } = this.props
+    decrementCount()
   }
 
   reset = () => {
-    const { dispatch } = this.props
-    dispatch(resetCount())
+    const { resetCount } = this.props
+    resetCount()
   }
 
   render () {
@@ -37,5 +38,6 @@ function mapStateToProps (state) {
   const { count } = state
   return { count }
 }
+const mapDispatchToProps = dispatch => bindActionCreators({ incrementCount, decrementCount, resetCount }, dispatch)
 
-export default connect(mapStateToProps)(Counter)
+export default connect(mapStateToProps, mapDispatchToProps)(Counter)
