+++ packages/next/README.md
@@ -2187,15 +2187,16 @@ Yes! Here's an example with [Apollo](/examples/with-apollo).
 <details>
 <summary>Can I use it with Redux and thunk?</summary>
 
-Yes! Here's an [example](/examples/with-redux)
+Yes! Here's an [example](/examples/with-redux-thunk)
 
 </details>
 
+<p></p>
 
 <details>
 <summary>Can I use it with Redux?</summary>
 
-Yes! Here's an [example](/examples/with-redux-no-thunk)
+Yes! Here's an [example](/examples/with-redux)
 
 </details>
